Add UDA to export .project.name
Since the default filter for some spurious reason does a regexp replace
on the project name, I at least want to export it to use it in my own
`project_template`.

This is only the top-level project, though, as the parent/child
relationship for projects is done outside of the issue object.